+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASE] - XXXX.XX.XX
+
+### Added
+
+- pyproject.toml: added dearpygui dependency to new viz_tools optional dependency group
+- runner.py: real-time plots for visualizing scenario data (e.g. relative distances)
+
 ## [v0.12.3] - 2025.10.06
 
 ### Fixed